feat(commitment-tree): consolidate server and storage into single server feature
Replace the separate server/storage features with a single server
feature that activates grovedb-storage and grovedb-bulk-append-tree/storage.
Set default-features = false on grovedb-bulk-append-tree dep so it
doesn't implicitly pull in RocksDB.
